reference to statement.subject and .object without revision
In the SpecIF 0.10.x branch there are no revisions (SpecIF "Level-1"). Since identifiers of all entities must be unique, a reference with specified revision for statement.subject and .object must point to the supplied entitiy to which the revision is constrained.
Note that entities with several revisions can be supplied and references to specific revisions can be made in SpecIF 0.11.1 and later.